requirements.txt
chromadb
fastapi
langchain
openai
pydantic
python-dotenv
python-multipart
tiktoken
uvicorn

streamlit==1.22.0
streamlit-chat==0.0.2.2
pdfplumber
matplotlib
pandasai==0.2.15
faiss-cpu

loguru
bs4
pypdf
openpyxl
#hashlib